Move new timezone_test.py tests over to test_timezone.py tests
